pyproject.toml
[project]
name = "nurtura"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ipdb>=0.13.13",
    "ipykernel>=6.31.0",
    "livekit-agents[baseten]>=1.1.7",
    "matplotlib>=3.9.4",
    "mediapipe>=0.10.14",
    "numpy>=2.0.2",
    "opencv-python>=4.11.0.86",
    "pygame>=2.6.1",
    "pyqt6>=6.10.0",
    "requests>=2.32.5",
    "scipy>=1.13.1",
    "torch>=2.8.0",
    "vapi-server-sdk>=1.7.3",
]
